+++ backend/src/HorseRacing.Domain/Entities/Compliance/RaceViolation.cs
@@ -1,9 +1,11 @@
+using HorseRacing.Domain.Entities.Tournaments;
+
 namespace HorseRacing.Domain.Entities;
 
 public class RaceViolation
 {
     public int Id { get; set; }
-    public int RaceId { get; set; }
+    public long RaceId { get; set; }
     public Race? Race { get; set; }
     public string Description { get; set; } = string.Empty;
     public string Penalty { get; set; } = string.Empty;
