+++ backend/src/HorseRacing.Domain/Entities/Tournaments/RaceResult.cs
@@ -3,6 +3,6 @@ namespace HorseRacing.Domain.Entities;
 public class RaceResult
 {
     public int Id { get; set; }
-    public int RaceId { get; set; }
+    public long RaceId { get; set; }
     public string Winner { get; set; } = string.Empty;
 }
